Expanded testing and fixed bugs found while testing

+++ WorkOut/app/src/main/java/Model/Location.java
@@ -7,32 +7,32 @@ import java.util.Objects;
 
 public class Location implements Parcelable
 {
-    private long longitude;
-    private long latitude;
+    private double longitude;
+    private double latitude;
 
     public Location() {}
-    public Location(long longitude, long latitude)
+    public Location(double longitude, double latitude)
     {
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public long getLatitude()
+    public double getLatitude()
     {
         return latitude;
     }
 
-    public long getLongitude()
+    public double getLongitude()
     {
         return longitude;
     }
 
-    public void setLatitude(long latitude)
+    public void setLatitude(double latitude)
     {
         this.latitude = latitude;
     }
 
-    public void setLongitude(long longitude)
+    public void setLongitude(double longitude)
     {
         this.longitude = longitude;
     }
@@ -61,8 +61,8 @@ public class Location implements Parcelable
     @Override
     public void writeToParcel(Parcel dest, int flags)
     {
-        dest.writeLong(longitude);
-        dest.writeLong(latitude);
+        dest.writeDouble(longitude);
+        dest.writeDouble(latitude);
     }
 
     @Override
